pyproject.toml
[project]
name = "rag-llm-chaining"
version = "0.1.0"
description = "Korean Legal RAG System with LLM Chaining"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "lanco", email = "lanco.gh@gmail.com"}
]
keywords = ["rag", "llm", "langchain", "korean", "legal"]

dependencies = [
    # LLM & Inference
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.27.0",
    "bitsandbytes>=0.42.0",
    "vllm>=0.4.0",

    # Embedding & Vector DB
    "sentence-transformers>=2.5.0",
    "chromadb>=0.4.22",
    "faiss-cpu>=1.7.4",

    # RAG Framework
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-huggingface>=0.0.3",

    # API & Web
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "gradio>=4.15.0",

    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "datasets>=2.17.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "tqdm>=4.66.0",
    "pyyaml>=6.0.1",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]
eval = [
    "ragas>=0.1.0",
    "deepeval>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
